Emulate Moab/Maui MAXIJOB via bf_max_job_user
Currently Slurm has bf_max_job_user=5 so we only look at any users
first 5 queued jobs when doing the backfill cycle. However, unlike
Moab/Maui, jobs such as that are not marked as blocked in Slurm
which causes a lot of user confusion.

We emulate this in showq by counting the number of waiting jobs of
a user and only print the first 5 waiting jobs and then the rest are
shown in the "blocked" section instead - which is hidden by default.
The user will see their own blocked jobs though of course.

Bump the version number to 0.12 in the process.