Swap to using Slurm's own reason codes for jobs.
With the proliferation of reason codes in Slurm it's infeasible
to track them all adn in the current code this just results in
messages like:

[Warn]: unknown job state - pending job reason = 20

With no indication of the job ID at all.

Instead we drop all that logic and call into Slurm via their C
function slurm_job_reason_string() to return the reason as a
string.

We then append a message referring the user to the VLSCI local
"explain" script whcih will pull the reason out of the Slurm
documentation.