Refactor the tests for a blocked job
Currently the code tests a list of reason codes in two places to
work out whether a job should be classed as blocked.  This means:
1) playing whack-a-mole chasing new Slurm reason codes
2) ensuring that each code is tested in both locations

The fix is to refactor the test out into a separate function that
returns 'true' for jobs that are blocked and 'false' for those that
aren't.  It does that by only testing for reason codes that indicate
that the job isn't blocked (of which there are few).

This means that we're more likely to be correct in future when new
codes are introduced, compared to the previous code.