When calculating time remaining take suspend time into account.
We hade a bug where a job that was suspended could generate a
warning that it was still running having overrun its timelimit.
This was because there was no login to take that into account.

Now (taking inspiration from Slurm's squeue) we do that so that
we don't generate bogus warnings. Validated using a job that has
been suspended on our cluster already.